Deprecate

The following types and methods are used in some old implementations, which may be deprecated in a future version.

FiniteMPS.LanczosInfo — Type

 struct LanczosInfo
      converged::Bool
      normres::Vector{Float64}
      numiter::Int64
      numops::Int64
      a::Vector{Float64}
      b::Vector{Float64}
 end

Similar to KrylovKit.ConvergenceInfo but delete residuals to save memory.

FiniteMPS.SparseProjectiveHamiltonian — Type

 struct SparseProjectiveHamiltonian{N} <: AbstractProjectiveHamiltonian  
	  El::SparseLeftTensor
	  Er::SparseRightTensor
	  H::NTuple{N, SparseMPOTensor}
	  si::Vector{Int64}
	  validIdx::Vector{Tuple}
	  E₀::Float64
 end

N-site projective Hamiltonian, sparse version. Note we shift H to H - E₀ to avoid numerical overflow.

Convention: – – – – | | | | | | | El– i – H1 – j –Er El– i – H1 – j – H2 – k –Er ... | | | | | | | – – – –

validIdx stores all tuples (i, j, ...) which are valid, i.e. all El[i], H1[i, j] and Er[j] are not nothing (N == 1).

FiniteMPS.ProjHam — Function

 ProjHam(Env::SparseEnvironment, siL::Int64 [, siR::Int64 = siL]; E₀::Number = 0.0)

Generic constructor for N-site projective Hamiltonian, where N = siR - siL + 1.

 ProjHam(Env::SimpleEnvironment, siL::Int64 [, siR::Int64 = siL])

Construct the special IdentityProjectiveHamiltonian from a simple environment.

FiniteMPS.PreFuseProjectiveHamiltonian — Type

 struct PreFuseProjectiveHamiltonian{N, Tl, Tr} <: AbstractProjectiveHamiltonian 
      El::Tl
      Er::Tr
      si::Vector{Int64}
      E₀::Float64
 end

Prefused N-site projective Hamiltonian. Note El and Er can be a original environment tensor or a prefused one, depending on N. If N == 1, only one of them will be prefused.

FiniteMPS.action0 — Function

 action0(obj::SparseProjectiveHamiltonian{0}, x::MPSTensor{2}; kwargs...) -> ::MPSTensor

Action of 0-site projective Hamiltonian on the rank-2 bond local tensors.

FiniteMPS.action1 — Function

 action1(obj::SparseProjectiveHamiltonian{1}, x::MPSTensor; kwargs...) -> ::MPSTensor

Action of 1-site projective Hamiltonian on the 1-site local tensors.

FiniteMPS.action2 — Function

 action2(obj::SparseProjectiveHamiltonian{2}, x::CompositeMPSTensor{2, T}; kwargs...) -> ::CompositeMPSTensor{2, T}

Action of 2-site projective Hamiltonian on the 2-site local tensors, wrapped by CompositeMPSTensor{2, T} where T<:NTuple{2,MPSTensor}.

 action2(obj::IdentityProjectiveHamiltonian{2}, x::CompositeMPSTensor{2, T}; kwargs...) -> ::CompositeMPSTensor{2, T}

Special case for IdentityProjectiveHamiltonian.